+++ PSFramework/functions/runspace/Write-PSFRunspaceQueue.ps1
@@ -18,6 +18,10 @@
 	.PARAMETER BulkValues
 		Write multiple values as separate entries.
 
+	.PARAMETER UseCurrent
+		Write to a queue in the current runspace workflow.
+		Only valid when used from within the code of a Runspace Workflow worker.
+
 	.PARAMETER Close
 		Closes the queue after writing the input.
 		This prevents further data to be added to the queue,
@@ -45,6 +49,7 @@
 		$Name,
 
 		[Parameter(Mandatory = $true, ValueFromPipeline = $true, ParameterSetName = 'Single')]
+		[Parameter(Mandatory = $true, ValueFromPipeline = $true, ParameterSetName = 'SingleCurrent')]
 		[AllowNull()]
 		$Value,
 
@@ -53,15 +58,21 @@
 		[object[]]
 		$BulkValues,
 
+		[Parameter(Mandatory = $true, ParameterSetName = 'SingleCurrent')]
+		[switch]
+		$UseCurrent,
+
 		[switch]
 		$Close,
 
-		[Parameter(ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true)]
+		[Parameter(ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'Single')]
+		[Parameter(ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'Multi')]
 		[PsfArgumentCompleter('PSFramework-runspace-workflow-name')]
 		[string[]]
 		$WorkflowName,
 
-		[Parameter(ValueFromPipeline = $true)]
+		[Parameter(ValueFromPipeline = $true, ParameterSetName = 'Single')]
+		[Parameter(ValueFromPipeline = $true, ParameterSetName = 'Multi')]
 		[PSFramework.Runspace.RSWorkflow[]]
 		$InputObject
 	)