+++ PSFramework/changelog.md
@@ -1,8 +1,30 @@
 ﻿# CHANGELOG
 
-## Unreleased
+## 1.11.341 (2024-07-08)
 
-- New: Configuration Validation: guid - ensures only legal guids can be added
+- New: Command New-PSFHashtable - returns a PsfHashtable object, a hashtable with default value option.
+- New: Configuration Validation: guid - ensures only legal guids can be added.
+- New: Type PsfHashtable - a hashtable that can have a default value.
+- New: Type Object.ObjectHost - added methods to modify members of a PSObject.
+- New: ParameterClass Certificateparameter - Resolves input as certificate with a private key.
+- New: ParameterClass PublicCertificateparameter - Resolves input as certificate.
+- Upd: New-PSFSupportPackage - Added a parameter to create the debug dump in a managed folder specific to the provided task name.
+- Upd: New-PSFSupportPackage - Added parameter to force-create parent folder of output path.
+- Upd: ConvertTo-PSFHashtable - Added parameter `-AsPsfHashtable` to return a PsfHashtable instead of a default Hashtable.
+- Upd: Runspace Workflows - Worker runspaces are now named "PSF-<Workflow>-<Worker>-<Index>".
+- Upd: Runspace Workflows - Errors that happened during a worker's processing now include the object being processed.
+- Upd: Runspace Workflows - Workflow & Worker object now expose information about their associated runspaces.
+- Upd: Add-PSFRunspaceWorker - Added parameter `-NoOutput` to ignore all worker code output.
+- Upd: Read-PSFRunspaceQueue - Added `-Peek` parameter to allow reading items without removing them from the queue.
+- Upd: Write-PSFRunspaceQueue - Added `-UseCurrent` parameter, to force autodetection of local runspace.
+- Upd: Wait-PSFRunspaceWorkflow - Added option to wait based on how long ago an item was added to a specified queue.
+- Upd: Type Utility.UtilityHost - added SetPrivateField method, using reflection to update a non-public field.
+- Fix: Read-PSFRunspaceQueue - Queue is not cleared when piping result to Select-Object -First X (#621).
+- Fix: Import during JEA session establishment fails - "Cannot bind empty value to Path".
+- Fix: Import-PSFpowerShellDataFile - Safe mode incorrectly reports "File is not safe to execute" on ArrayLiteralAsts inside of a psd1 file.
+- Fix: New-PSFMessageLevelModifier - Is case sensitive when comparing function names.
+- Fix: Runspace Workflows - Do not properly dispose runspaces after processing completes. (#619)
+- Fix: Runspace Workflows - Will add an empty output in a situation where a worker should not produce output in the traditional sense.
 
 ## 1.10.318 (2023-11-10)